# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "GatewaySummary": {
        "additionalProperties": false,
        "description": "Per-client totals over a time window, most expensive first.",
        "properties": {
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "clients": {
            "items": {
              "$ref": "#/components/schemas/GatewayClientTotals"
            },
            "title": "Clients",
            "type": "array"
          }
        },
        "required": [
          "clients",
          "as_of"
        ],
        "title": "GatewaySummary",
        "type": "object"
      },
...
      "GatewayTurnList": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "description": "Server time; pass it as `since` on the next live poll.",
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "turns": {
            "description": "Newest first.",
            "items": {
              "$ref": "#/components/schemas/GatewayTurn"
            },
            "title": "Turns",
            "type": "array"
          }
        },
        "required": [
          "turns",
          "next_cursor",
          "as_of"
        ],
        "title": "GatewayTurnList",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestAccepted": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "description": "Spans queued for writing (always the whole batch).",
            "title": "Accepted",
            "type": "integer"
          }
        },
        "required": [
          "accepted"
        ],
        "title": "IngestAccepted",
        "type": "object"
      },
...
      "SpanBatch": {
        "additionalProperties": false,
        "description": "Body of ``POST /v1/spans``.",
        "properties": {
          "spans": {
            "items": {
              "$ref": "#/components/schemas/Span"
            },
            "maxItems": 1000,
            "minItems": 1,
            "title": "Spans",
            "type": "array"
          }
        },
        "required": [
          "spans"
        ],
        "title": "SpanBatch",
        "type": "object"
      },
...
      "TraceDetail": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "description": "Server time of this response; pass it as `since` on the next live poll.",
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "spans": {
            "description": "Spans ordered by start_time: all of them, or with `since` only those stored after it.",
            "items": {
              "$ref": "#/components/schemas/Span"
            },
            "title": "Spans",
            "type": "array"
          },
          "trace": {
            "$ref": "#/components/schemas/TraceSummary"
          }
        },
        "required": [
          "trace",
          "spans",
          "as_of"
        ],
        "title": "TraceDetail",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LucentPad API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/gateway/anthropic/{path}": {
      "get": {
        "description": "Set `ANTHROPIC_BASE_URL` (Claude Code) or `COPILOT_PROVIDER_BASE_URL` with `COPILOT_PROVIDER_TYPE=anthropic` (Copilot CLI) to `<host>/gateway/anthropic`. Forwards `v1/messages`, `v1/messages/count_tokens` and `v1/models`; bodies and streams pass through unchanged. The client's key is forwarded, never stored.",
        "operationId": "gateway_anthropic_get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anthropic API, proxied (Claude Code, Copilot CLI)",
        "tags": [
          "gateway"
        ]
      },
      "post": {
        "description": "Set `ANTHROPIC_BASE_URL` (Claude Code) or `COPILOT_PROVIDER_BASE_URL` with `COPILOT_PROVIDER_TYPE=anthropic` (Copilot CLI) to `<host>/gateway/anthropic`. Forwards `v1/messages`, `v1/messages/count_tokens` and `v1/models`; bodies and streams pass through unchanged. The client's key is forwarded, never stored.",
        "operationId": "gateway_anthropic_post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Anthropic API, proxied (Claude Code, Copilot CLI)",
        "tags": [
          "gateway"
        ]
      }
    },
    "/gateway/openai/{path}": {
      "get": {
        "description": "Use `<host>/gateway/openai/v1` as the OpenAI-compatible base URL. Forwards `v1/chat/completions` and `v1/models`; bodies and streams pass through unchanged.",
        "operationId": "gateway_openai_get",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "OpenAI API, proxied (Copilot Chat Custom Endpoint, Copilot CLI)",
        "tags": [
          "gateway"
        ]
      },
      "post": {
        "description": "Use `<host>/gateway/openai/v1` as the OpenAI-compatible base URL. Forwards `v1/chat/completions` and `v1/models`; bodies and streams pass through unchanged.",
        "operationId": "gateway_openai_post",
        "parameters": [
          {
            "in": "path",
            "name": "path",
            "required": true,
            "schema": {
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "OpenAI API, proxied (Copilot Chat Custom Endpoint, Copilot CLI)",
        "tags": [
          "gateway"
        ]
      }
    },
...
    "/v1/gateway/summary": {
      "get": {
        "description": "Per-client sessions, turns, tokens and cost over the window.",
        "operationId": "gateway_summary_v1_gateway_summary_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewaySummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not implemented yet."
          }
        },
        "summary": "Gateway Summary",
        "tags": [
          "query"
        ]
      }
    },
    "/v1/gateway/turns": {
      "get": {
        "description": "Gateway model calls, newest first.",
        "operationId": "gateway_turns_v1_gateway_turns_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "description": "Repeat for OR.",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeat for OR.",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "description": "Live polling: turns stored after this time (previous `as_of`).",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Live polling: turns stored after this time (previous `as_of`).",
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayTurnList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not implemented yet."
          }
        },
        "summary": "Gateway Turns",
        "tags": [
          "query"
        ]
      }
    },
...
    "/v1/spans": {
      "post": {
        "description": "Queue a batch for writing and return at once; spans are readable shortly after.",
        "operationId": "ingest_spans_v1_spans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpanBatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Body larger than the ingest limit."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Ingest queue full; nothing from the batch was accepted. Retry after the `Retry-After` header (seconds).",
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          }
        },
        "summary": "Ingest Spans",
        "tags": [
          "ingest"
        ]
      }
    },
...
    "/v1/traces/{trace_id}": {
      "get": {
        "operationId": "get_trace_v1_traces__trace_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "trace_id",
            "required": true,
            "schema": {
              "title": "Trace Id",
              "type": "string"
            }
          },
          {
            "description": "Live polling: only spans stored after this time (the previous response's `as_of`), with a few seconds of overlap; merge by `span_id`. The summary is always current.",
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Live polling: only spans stored after this time (the previous response's `as_of`), with a few seconds of overlap; merge by `span_id`. The summary is always current.",
              "title": "Since"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TraceDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Trace",
        "tags": [
          "query"
        ]
      }
    }
  }
}